Rewrites the script using less dependencies.
Instead we use the power of Python and good idioms (such as not using
mutable global variables) in order to create a more maintainable and
easier to understand script. How is it easier to understand? Well, for
one you don't need to lookup what kind of a framework 'twisted' is or
understand why we need it, when we can just use time.sleep(x) which is
much more intuitive.